+++ src/main/java/com/promesa/promesa/domain/inquiry/api/InquiryController.java
@@ -24,30 +24,4 @@ public class InquiryController {
     public ResponseEntity<List<InquiryResponse>> getInquiriesByArtist(@RequestParam Long artistId) {
         return ResponseEntity.ok(inquiryService.getInquiriesByArtist(artistId));
     }
-
-    @Operation(summary = "문의 등록")
-    @PostMapping()
-    @PreAuthorize("hasAnyRole('ADMIN', 'ARTIST')")
-    public ResponseEntity<String> createInquiry (
-            @RequestBody @Valid AddInquiryRequest request
-    ) {
-        return ResponseEntity.ok(inquiryService.createInquiry(request));
-    }
-
-    @Operation(summary = "문의 삭제")
-    @DeleteMapping("/{inquiryId}")
-    @PreAuthorize("hasAnyRole('ADMIN', 'ARTIST')")
-    public ResponseEntity<String> deleteInquiry (@PathVariable Long inquiryId) {
-        return ResponseEntity.ok(inquiryService.deleteInquiry(inquiryId));
-    }
-
-    @Operation(summary = "문의 수정")
-    @PutMapping("/{inquiryId}")
-    @PreAuthorize("hasAnyRole('ADMIN', 'ARTIST')")
-    public ResponseEntity<String> updateInquiry (
-            @PathVariable Long inquiryId,
-            @RequestBody @Valid UpdateInquiryRequest request
-    ) {
-        return ResponseEntity.ok(inquiryService.updateInquiry(inquiryId, request));
-    }
 }
